backend/requirements.txt
# Core Voice Agent Dependencies
aiohttp
python-dotenv
SpeechRecognition
PyAudio
edge-tts
pydub
simpleaudio

# For Conversation History
supabase
sentence-transformers
numpy
psycopg2-binary

# For Web Server
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
